added refresh support
