+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/common/config/SubmissionConfig.kt
@@ -28,14 +28,12 @@ class SubmissionConfig(
     @Suppress("LongParameterList")
     fun submissionService(
         subRepository: SubmissionRepository,
-        persistenceService: SubmissionPersistenceService,
         serializationService: SerializationService,
         userPrivilegeService: IUserPrivilegesService,
         queryService: SubmissionQueryService,
         submissionSubmitter: SubmissionSubmitter
     ): SubmissionService = SubmissionService(
         subRepository,
-        persistenceService,
         serializationService,
         userPrivilegeService,
         queryService,
