+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/submission/web/handlers/SubmitWebHandler.kt
@@ -3,12 +3,15 @@ package ac.uk.ebi.biostd.submission.web.handlers
 import ac.uk.ebi.biostd.files.service.UserFilesService
 import ac.uk.ebi.biostd.integration.SerializationService
 import ac.uk.ebi.biostd.integration.SubFormat
+import ac.uk.ebi.biostd.persistence.integration.FileMode
 import ac.uk.ebi.biostd.submission.domain.helpers.RequestSources
 import ac.uk.ebi.biostd.submission.domain.helpers.SourceGenerator
 import ac.uk.ebi.biostd.submission.domain.service.SubmissionService
 import ac.uk.ebi.biostd.submission.model.SubmissionRequest
 import ac.uk.ebi.biostd.submission.web.model.ContentSubmitWebRequest
 import ac.uk.ebi.biostd.submission.web.model.FileSubmitWebRequest
+import ac.uk.ebi.biostd.submission.web.model.RefreshWebRequest
+import ebi.ac.uk.extended.mapping.serialization.to.toSimpleSubmission
 import ebi.ac.uk.io.sources.FilesSource
 import ebi.ac.uk.model.Submission
 import ebi.ac.uk.model.SubmissionMethod.FILE
@@ -30,6 +33,12 @@ class SubmitWebHandler(
         val source = sources(request.user, sub, request.files)
         val submission = withAttributes(submission(request.submission, request.format, source), request.attrs)
         return submissionService.submit(SubmissionRequest(submission, request.user, source, PAGE_TAB, request.fileMode))
+    }
+
+    fun refreshSubmission(request: RefreshWebRequest): Submission {
+        val submission = submissionService.getSubmission(request.accNo).toSimpleSubmission()
+        val source = sources(request.user, submission)
+        return submissionService.submit(SubmissionRequest(submission, request.user, source, PAGE_TAB, FileMode.MOVE))
     }
 
     fun submit(request: FileSubmitWebRequest): Submission {
