+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/submission/web/resources/SubmissionResource.kt
@@ -20,7 +20,6 @@ import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.ModelAttribute
 import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.ResponseBody
 import org.springframework.web.bind.annotation.RestController
@@ -91,22 +90,6 @@ class SubmissionResource(
         @ApiParam(name = "accNo", value = "The accession number of the submission to be deleted")
         @PathVariable accNo: String
     ): Unit = submissionsWebHandler.deleteSubmission(accNo, user)
-
-    @PostMapping("refresh/{accNo}")
-    @ApiOperation("Update submission based on system db stored information")
-    @ApiImplicitParam(
-        name = "X-Session-Token",
-        value = "User authentication token",
-        required = true,
-        paramType = "header")
-    fun refreshSubmission(
-        @BioUser user: SecurityUser,
-
-        @ApiParam(name = "accNo", value = "The accession number of the submission to be refresh")
-        @PathVariable accNo: String
-    ) {
-        submissionsWebHandler.refreshSubmission(accNo, user)
-    }
 
     private fun SimpleSubmission.asDto() =
         SubmissionDto(accNo, title.orEmpty(), version, creationTime, modificationTime, releaseTime, method)
