+++ submission/submission-webapp/src/main/kotlin/ac/uk/ebi/biostd/submission/web/resources/SubmitResource.kt
@@ -7,6 +7,7 @@ import ac.uk.ebi.biostd.persistence.integration.FileMode
 import ac.uk.ebi.biostd.submission.converters.BioUser
 import ac.uk.ebi.biostd.submission.web.handlers.SubmitWebHandler
 import ac.uk.ebi.biostd.submission.web.model.ContentSubmitWebRequest
+import ac.uk.ebi.biostd.submission.web.model.RefreshWebRequest
 import ebi.ac.uk.model.Submission
 import ebi.ac.uk.model.constants.APPLICATION_JSON
 import ebi.ac.uk.model.constants.ATTRIBUTES
@@ -21,6 +22,7 @@ import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
 import org.springframework.http.MediaType.APPLICATION_JSON_VALUE
 import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
@@ -98,5 +100,21 @@ class SubmitResource(private val submitWebHandler: SubmitWebHandler) {
     ): Submission {
         val request = ContentSubmitWebRequest(submission, user, JSON_PRETTY, mode, attributes.orEmpty(), emptyList())
         return submitWebHandler.submit(request)
+    }
+
+    @PostMapping("refresh/{accNo}")
+    @ApiOperation("Update submission based on system db stored information")
+    @ApiImplicitParam(
+        name = "X-Session-Token",
+        value = "User authentication token",
+        required = true,
+        paramType = "header")
+    fun refreshSubmission(
+        @BioUser user: SecurityUser,
+
+        @ApiParam(name = "accNo", value = "The accession number of the submission to be refresh")
+        @PathVariable accNo: String
+    ) {
+        submitWebHandler.refreshSubmission(RefreshWebRequest(accNo, user))
     }
 }
